Fix error from clang unsigned overflow sanitizer.
This is actually fully defined behavior, but it still makes sense to cast here.

+++ pb_decode.c
@@ -1546,13 +1546,13 @@ static bool checkreturn pb_dec_string(pb_istream_t *stream, const pb_field_iter_
 
     if (!pb_decode_varint32(stream, &size))
         return false;
-    
+
     /* Space for null terminator */
-    alloc_size = (size_t)(size + 1);
-    
+    alloc_size = (size_t)((size_t)size + 1);
+
     if (alloc_size < size)
         PB_RETURN_ERROR(stream, "size too large");
-    
+
     if (PB_ATYPE(field->type) == PB_ATYPE_POINTER)
     {
 #ifndef PB_ENABLE_MALLOC
